move model and mixin folders
