+++ frigate/data_processing/common/license_plate/mixin.py
