+++ frigate/data_processing/common/license_plate/model.py
@@ -5,7 +5,7 @@ from frigate.embeddings.onnx.lpr_embedding import (
     PaddleOCRRecognition,
 )
 
-from ..types import DataProcessorModelRunner
+from ...types import DataProcessorModelRunner
 
 
 class LicensePlateModelRunner(DataProcessorModelRunner):
