+++ frigate/data_processing/real_time/license_plate.py
@@ -6,10 +6,12 @@ import logging
 import numpy as np
 
 from frigate.config import FrigateConfig
-from frigate.data_processing.common.license_plate import (
+from frigate.data_processing.common.license_plate.mixin import (
     LicensePlateProcessingMixin,
 )
-from frigate.data_processing.common.license_plate_model import LicensePlateModelRunner
+from frigate.data_processing.common.license_plate.model import (
+    LicensePlateModelRunner,
+)
 
 from ..types import DataProcessorMetrics
 from .api import RealTimeProcessorApi
